Add files via upload
Further fixes to the Predicted FEV1 work - now using a calculated Age. 
Also revisited the duplicate handling. For Lung Function measures, Andres felt the triplet measures in <30min window was mirroring clinical procedure to take the best of 3 readings. So for Lung Function measures, I changed the handling of dupes to take the max of those within a 30min window. For other non-Activity measures, I left as the chronologically last row.
- added some fixes to anomalous data
- refactored the code for plotting the graphs to more easily allow different slices of patients to be run.